Replace the stm32_flash device with a standard memory alias
When I originally wrote this code, I created a new device just
to perform the memory aliasing.  Later I found out that QEMU has
this capability built-in and only requires a couple of lines of
code to set up.  I am now redoing it the right way.

Note that the stm32_p103_demo's that I am using for testing
mostly work without having the aliased memory.  The only issue I
have seen is when using interrupts.  The STM32 standard libraries
set the interrupt vector table base address (in the System Control
VTOR register) to 0x08000000 rather than 0x00000000, which
results in a hang when the interrupt is called.