stm32_dma: update IRQs based on ISR bits
Previous, the DMA IRQ was always fired, so long as interrupts were
enabled, rather than only when it was necessary.

Instead, for each DMA stream, explicitly check if any of the relevant DMA
LISR/HISR bits are set, indicating that an interrupt should be active.
If so, raise the IRQ in qemu.  If none of the bits are set, lower the IRQ
in qemu.

When the LIFCR or HIFCR registers are written, update all IRQs, as they
may no longer need to be asserted.